[Claude code] Add a '--lrc' arg to the play command.
I used this prompt:

In src/main.rs, the 'play' command takes a music file and loads both the
file and an LRC file whose filename is determined by replacing the music
file's extension with ".lrc".  In addition to this, I'd like you to add
an optional '--lrc=<filename>' argument that allows the user to specify
an LRC file explicitly. If this argument isn't specified, the default
(current) behavior should be used.